airflow_astro_docker/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
kaggle==1.6.17

apache-airflow-providers-dbt-cloud